Add a basic integration of the block email editor to the woo core (#55754)
* Add email editor php package to composer.json

* Add basic email editor integration files

The integration will only be enabled when the block_email_editor feature flag is set

* Add the pnpm build commands for the email editor js package

I would ideally like to ensure we use the same build system and process as the rest of the monorepo packages but that will increase the scope of this ticket.

Adding webpack config will be done in https://github.com/woocommerce/woocommerce/issues/55688#issue-2866084770

* Fix errors with the PersonalizationTagManager and typo with the WooMailContentPattern

* Add JS integration support for the email editor package

* Add basic integration of the block email editor to WooCommerce core

This commit adds the initial integration of the block email editor to WooCommerce core. The changes include:

- Registering the email editor post type
- Adding personalization tags for WooCommerce emails
- Implementing a basic email template
- Creating patterns for email content
- Setting up service providers and integration hooks

The feature is experimental and not ready for public use. It will only be active when the block_email_editor feature flag is enabled.

* Refactor email editor namespaces and add email patterns and templates

This commit updates the email editor integration by:
- Renaming namespaces for patterns and templates to be more descriptive

The changes improve the organization of the email editor components and provide initial patterns and templates for WooCommerce emails.

* Update email editor integration with naming and dependency improvements (use woo_email over woo_mail)

This commit refines the WooCommerce email editor integration by:
- Updating naming conventions for email-related classes and post types
- Adding dependency checks for the email editor
- Adding a temporary comment for the save button label integration

* Refine save button logic and update WooCommerce email template title

- Adjust save button disabled state to only disable when no edits are present
- Update WooCommerce email template title to use proper capitalization