Add/email editor sending transactional emails (#55940)
* Add placeholder block for Woo generated content

This commit adds very basic block that act as a placeholder for
the Woo generated content for the transactional emails (order details, shipping address, etc.).

It is registered only in the email editor via JS so that the block is not globally available.
The idea is that we will extend this block to preview Woo Content in the editor per email type.

* Add Woo Content placeholder block to the email content pattern

This ensures that when creating a new email with the pattern it
contains the Woo Content block.
The block is locked so that a user can't remove it.

* Add Woo content block render and register it to email renderer

The block will render a placehoder that will be replaced
by the real content during sending

* Use block email for transactional email when defined.

When the feature is active and there is an email created in the block
editor and associated with the transactional email, we render the block email
using MailPoet\EmailEditor\Engine\Renderer\Renderer.
We grab the main content of Woo email and place it instead of the placeholder
generated by Woo Content block.

To grab only the main content, we need to unregister filters for the email header and footer.
We then restore the filters to be able to correctly send other emails that might not have
an associated email created in the block editor..

* Add personalized tags processing when sending transactional email

* Add more context data for personalization from WC_Email

* Add changelog

* Skip block email rendering test if requirements are not fulfilled

The test may run on older WP versions and we need to skip to prevent errors.

* Move BlockEmailRenderer filter registration to init

The init method is called automatically when the service is
created in the container. We can use this to save some lines of code.

* Add select with email types to the block email editor sidebar.

This commit allows users to select which email type will be overridden
by the block-based email edited in the editor.
The association is currently saved in the post.slug ensuring that
There will be only one block email per email type.

* Update email editor code namespaces and prefixes after rebase

* Update query for block based emails to use slug aka post_name

* When there is a block-based email, switch the plain email type to HTML