Update Block Templates Package for React 18 (#56341)
* Update register-woo-block-type.ts for React 18

These changes should resolve the TypeScript error while maintaining
the functionality of the code. The approach uses `any` in a
controlled way with appropriate ESLint comments to acknowledge the
type safety compromise, which is necessary given the comment about
not being able to include proper type definitions.

* Update `select` param type in prod editor pkg

Changed the `select` parameter type in `useEvaluationContext`
from `typeof WPSelect` to `any` with appropriate ESLint comments.

* Remove unused arg from getEvaluationContext

The `select` arg is not used in the getEvaluationContext method. This
is removed here, per [this suggestion](https://github.com/woocommerce/woocommerce/pull/56341#discussion_r1993495077).

* Add changefile(s) from automation for the following project(s): @woocommerce/product-editor, @woocommerce/block-templates

* Use generic `SelectType` instead of `any`

Since we've already defined the generic `SelectType`, this should be
used instead of `any` in the `useSelect` callback arg.

* Add `SelectType` generic to use instead of `any`

Replicate the `SelectType` generic we have defined in
`register-woo-block-type.ts` so we can use here instead of `any` as
well.

* Remove unneeded  comments

* Add type assertion to select('core') return value

This adds a type assertion to properly type the return value
from `select('core')` by creating a coreStore variable with the
correct type signature for the `getEditedEntityRecord` method.

Additionally, it adds a type assertions for postType to explicitly
tell TypeScript that it should be treated as a string when passed
to both `useEntityId` and `getEditedEntityRecord`.

* Remove unneeded comments

---------

Co-authored-by: github-actions <github-actions@github.com>
Co-authored-by: roykho <roykho77@gmail.com>