[Email Editor] Improve Woo content styles in block email (#56199)
* Alter styles for Woo emails when created via block editor

In the flow in WC_Email the content is generated and after styles are inlined.
We replace the email content when it is generated, but we need to provide styles
for inlining, which is done later in the flow.
We can apply only the block editor CSS or combine the editor's CSS with Woo email
CSS. I chose to combine because Woo email CSS contains some styles for the content which the editor's CSS doesn't cover.
I removed font-family and color styles from the Woo CSS to keep a unified look of the content.

* Move the logic for capturing the Woo mail content to BlockEmailRenderer

This is the first step of refactoring the functionality for capturing
Woo email's content to an extra service.
We will need such a service to be able grab the content for block
or for preview.

* Refactor functionality for capturing content from WC_Email into a service

The service extracts the main content from WC_Emails and inlines
Styles are generated from settings in the block editor.
The intention is to reuse it, for example, in the preview.

* Add changelog

* Remove unused service from BlockEmailRenderer

* Update prefixes of block related filters and actions in email renderer

We need to distinguish general block filters from those related to email.
This commit changes prefix from `woocommerce_blocks` to `woocommerce_email_blocks`