Performance: Simpler query for orders months filter dropdown (#55510)
* Perf: Simpler query for orders months filter

This replaces the expensive query that determines every month that an
order was created in with a simple one that just takes the oldest order.
Then we build the list of options by adding every month from then until
the current one. The tradeoff is that if a store has some large gaps of
time between some orders, the dropdown may show months where no orders
were actually created.

Fixes #54579

* Appease the linter

* Add option to toggle which months method to use

Instead of always using the simplified query for the months filter,
which might trade worse UX for negligible performance gains for stores
with small, sporadic order volume, let's instead keep using the current
query method, and just add a toggle. This way the high volume stores can
enable the simple query as part of their tuning process, and other
stores won't see any change.

This also allows us to make use of the caching from the current method
when using the simplified method.

This also adds a unit test for the simplified method.

* Ensure retrieved option value is boolean

* Revert to original approach

This reverts 606dac73389209aff34435e8937db2c4e5b1fddf and
088ebea2dcbb11f7f36b88a92e6b160f6e825b40 to return to the original
approach of only offering the simplified months filter options and
deprecating the method that runs the expensive query and caches it.

This additionally fixes up the unit tests to work correctly with the
simplified approach.

* Appease the linter

* Account for timezone offset from GMT

Ensure that the month options reflect the months in the site's timezone.
The order date is stored in GMT, so it could end up being a different
month and/or year than the local order date.

Also updates the months query so that the trash status is included in
the SQL statement via the prepare method, rather than interpolation.

* Swap expected and actual on several tests