Improvements and fixes when recounting product category and tag terms (#56543)
* Ensure wc_change_term_counts handles multiple taxonomies

This fixes a subtle bug that can happen when you use get_terms and
specify multiple taxonomies in an array instead of a single taxonomy.

* Add some initial unit tests

* Appease the linter

* Ensure _update_post_term_count gets correct IDs

Fixes #44462

* Add $include_callback param to wc_recount_all_terms

This allows us to avoid running the standard _update_post_term_count
callback when it isn't necessary, which in turn avoids running a bunch
of extra queries.

* Avoid querying product_visibility term IDs more than once

* Revamp _wc_term_recount and improve doc block

* Add changelog file

* Attempt to avoid unit test static variable snafu