Add email listing linked to the email block editor into Settings > Email (#56641)
* Replace email notifications list in email settings with a react slotfill

We will render the DataViews listing component in that area.

WOOPLUG-3336

* Add basic email listing using Dataviews into email settings.

The source data for the listing is a list of emails provided
by WC()->mailer()->get_emails().
The list is then enriched with the woo_email posts data and passed to
the Dataview component.

WOOPLUG-3336

* Display email description in the listing table

WOOPLUG-3336

* Add email status to the email listing

WOOPLUG-3336

* Add preview action to emails listing

WOOPLUG-3336

* Add support for search and status filter to the email listing

WOOPLUG-3336

* Configure sorting in emails listing

WOOPLUG-3336

* Add support for pagination to email listing

WOOPLUG-3336

* Add dummy enable and disable action to the email listing

WOOPLUG-3336

* Add a component for displaying the recipient list in the email listing

WOOPLUG-3336

* Use Woo transactional email status instead of an email post status

The email post status is not relevant to the transactional email status.

WOOPLUG-3336

* Add dummy send test email action

WOOPLUG-3336

* Mark the actions that are not implemented yet as disabled

WOOPLUG-3336

* Update Email section description and move the SMTP help info below listing

WOOPLUG-3336

* Ensure that changes in the email listing view don't trigger settings change

WOOPLUG-3336

* Move email status logic to client

WOOPLUG-3336

* Prevent preview and edit action for emails that don't have associated post

WOOPLUG-3336

* Add functionality to toggle email status in listing in settings

Originally, I attempted to implement the status switching and its state fully on top
of woocommerce/data. This would require fetching settings for all email types, which
triggers a request per email type. To avoid multiple requests, I
decided to keep the local state inside the hook and fetch only a single setting
after a user toggles the state.

WOOPLUG-3336

* Add memoization and fix TS issues

WOOPLUG-3336

* Remove unnecessary Id field from email listing

WOOPLUG-3336

* Add @wordpress/dataviews package for email listing

WOOPLUG-3336

* Fix lint issues

WOOPLUG-3336

* Add changelog entry for email listing

WOOPLUG-3336

* Route edit link for emails that don't support block editor to current settings

When a transactional email doesn't support the block editor we route the
edit action to the old email settings screen.

WOOPLUG-3336

* Make the status update in the email listing optimistic

This commit improves UI feedback for slow connections.
The status is switched immediately, and in case it fails, we switch
it back.

WOOPLUG-3336

* Add e2e test for the new email listing in the settings

WOOPLUG-3336

* Update the email statuses in the email listing to enabled/disabled

* Use the correct section ID for links to the email manage page

* Use WCTransactionalEmailPostsManager for fetching email post id

This commit removes the duplicate functionality for fetching the post_id.