Fix: Prevent null offset warning in has_block_including_patterns method found in BlockTemplateUtils.php (#56981)
* Fix: Prevent null offset warning in BlockTemplateUtils.php

In the function `has_block_including_patterns`, a warning was being triggered:
"Trying to access array offset on value of type null" due to a lack of checks for a possibly null block pattern.

This commit adds validation to ensure that the block pattern exists and contains the specified content before attempting to parse it. This change improves stability and eliminates warnings when specific block patterns are not registered or when their content is not set.

* Add changefile(s) from automation for the following project(s): woocommerce

* Update plugins/woocommerce/src/Blocks/Utils/BlockTemplateUtils.php

modified code to pass lint tests.

Co-authored-by: Thomas Roberts <5656702+opr@users.noreply.github.com>

* Update plugins/woocommerce/src/Blocks/Utils/BlockTemplateUtils.php

Removed $pattern check as it is not necessary in this case and fixed formatting to pass lint tests

Co-authored-by: Thomas Roberts <5656702+opr@users.noreply.github.com>

---------

Co-authored-by: github-actions <github-actions@github.com>
Co-authored-by: Thomas Roberts <5656702+opr@users.noreply.github.com>