Update method for generating order months dropdown (#56879)
* Update method for generating order months dropdown

In #55510 we changed how the Months filter dropdown on the Orders list
table is generated so that it uses a much more efficient query. However,
it introduced a bug that could happen at the end of the month when the
following month had fewer days in it. This was surfaced via a unit test,
fortunately. In this PR we are making the method more robust so that the
number of days in the month doesn't matter, and also improving the
behavior during an edge case where all existing orders are in the future

* Appease the linter

* Fix linting: Spacing in prepare statement

* Fix flaky e2e test

* Simplify months query and option building algo

This does away with conditionally running two separate queries to get
the minimum and maximum order creation dates, in favor of one query that
selects the min and max and returns them as a single row. It also makes
some other changes to try and be consistent about the timezones of the
dates involved.

* Appease the linter

* Change method back to protected

---------

Co-authored-by: Tom Cafferkey <tjcafferkey@gmail.com>