Allow edit only the main email template (#57246)
* Update email editor page renderer also to support template

When a template is present in the query parameters, we ignore post
params that are present only to pass conditions on post.php, and
we setup the editor to use the template as the primary document.
WOOPLUG-3312

* Update the email editor to work only with a template

- template uses string IDs so we need to allow them
- we need a few adjustments in header so that save buttons make sense
WOOPLUG-3312

* Move email sectio description to react

This change will make it easier to add the edit template button.
WOOPLUG-3312

* Add Edit template button to the email settings page

We need to pass a post ID to get over checks in post.php. I used
post from one of the emails from the listing.
WOOPLUG-3312

* Add changelogs

WOOPLUG-3312

* Reintroduce the Save email button in template mode to return back to settings

WOOPLUG-3312

* Hide irrelevant preview options for template mode

WOOPLUG-3312