Migrate block email editor to the wordpress/editor Editor component (#57775)
* Add private APIs for the Editor component

* Replace the custom editor layout with the Editor component

* Move layout styles into the useEmailCss hook

The logic for restricting email width to the width defined in the layout settings
was in the VisualEditor component.

The VisualEditor is not used any more so I moved the logic to the hook,
that generates CSS for the editor content.

WOOPLUG-3951

* Fix PageRenderer to support creating new emails

Post ID might be 0 when we create a new email so we must use
Post ID from the newly created email post.
WOOPLUG-3951

* Reintroduce TemplateSelection functionality

WOOPLUG-3951

* Remove unnecessary patterns fetching

The patterns are fetched within the Editor component

WOOPLUG-3951

* Migrate email styles sidebar to PluginSidebar API

WOOPLUG-3951

* Add custom header to the email styles sidebar

The default PluginSidebar header contains a button with a star icon that
completely hides the sidebar even from the top editor bar.
We don't have such functionality in designs.

WOOPLUG-3951

* Update Navigator component in the email styles sidebar

Navigator is now non-experimental  component.
https://wordpress.github.io/gutenberg/?path=/docs/components-navigator--docs
WOOPLUG-3951

* Update Woo Email CPT labels

These labels are displayed in some places of the editor (e.g. Email
in the sidebar tabs). Woo Email is the internal name.

WOOPLUG-3951

* Add send preview to preview options

PluginPreviewMenuItem adds a new item in the preview toggle dropdown.
WOOPLUG-3951

* Re-enable fullscreen mode

The full-screen mode controls are not part of wordpress/editor.
We need to add more menu via ViewMoreMenuGroup component and control
its' state in our store.

Inspired by edit-post package.

WOOPLUG-3951

* Add autosave, error boundary, basic shortcuts, and snackbar notices

WOOPLUG-3951

* Hide original status panel

The original status pabel contains some options that don't work
for emails (e.g., user can switch to a site template), and we can't hide it selectively.

WOOPLUG-3951

* Add settings panel for emails

WOOPLUG-3951

* Add sender settings to template mode

We currently can't use PluginPostStatusInfo slot because post-status
panel is hidden.
We add the sender settings as an extra panel.
WOOPLUG-3951

* Introduce Command menu with basic commands

WOOPLUG-3951

* Remove dependency on @wordpress/interface

These actions are no longer needed as they are completely covered
by the Editor component.

WOOPLUG-3951

* Fix fetching an email not associated with transactional email

When $email_type was null it caused 500 because get_email_by_type expects string.
WOOPLUG-3951

* Add the Save Email button

I placed the code intentionally in the hack folder.
It is one of the things we should address upstream in the future.
Currently, we hide the original button when there is nothing
to save and we replace it with our custom button.

WOOPLUG-3951

* Configure autosave timeout and disable code editor

WOOPLUG-3951

* Add validation notices and custom snackbars

The custom snackbars are included in EditorNotices and contain
text replacement functionality to ensure notices are understandable in email context.

I added additional hacky slot for rendering validation notice into
the proper area.

WOOPLUG-3951

* Fix tiny CSS issues

- floating preview text lenght info was not counted to element height
- slightly missplaced icon in preview dropdown

* Fix global padding is not applied

This was previously working because se manually added
class 'has-global-padding' to the root container.
We no longer have access to add class to the container so
we need to generate CSS ourselves.

WOOPLUG-3951

* Fix loading CSS from the server into the editor

This was previously done in a low-level component we no longer use.

WOOPLUG-3951

* Remove the unwanted margin for the first content element

The editor has newly class wp-site-blocks added automatically and this rule
was conflicting how we set margins for block gap settings.

I also removed the comment because it was not a valid CSS comment.
WOOPLUG-3951

* Hide the huge title input

In the preview dropdown, there is an option to hide the template content.
In such case, the editor switches to post-only mode and displays the big
title input on top. Such a title makes no sense in an email context as it
is not part of the final render.

This commit adds CSS to hide it.
WOOPLUG-3951

* Fix useSelect warning about returning different values for template

The `useSelect` hook returns different values when called with the same state and parameters.
This can lead to unnecessary re-renders and performance issues if not fixed.

Non-equal value keys: template

WOOPLUG-3951

* Fix the useSelect warning about globalStylesPost

There was a strange fallback to template post when global styles post was not present.
After I removed the fallback I had to make a few fixes in style hooks.

Warning was:

The `useSelect` hook returns different values when called with the same state and parameters.
This can lead to unnecessary re-renders and performance issues if not fixed.

Non-equal value keys: globalStylesPost

WOOPLUG-3951

* Cleanup unused components from the email editor

These components are no longer used after we switched to the Editor
WOOPLUG-3951

* Cleanup SCSS files and fix CSS loading

The CSS loading issue was caused by content related CSS for richtext defined in CSS which is used for layout
and was added additionally to the iframe.

I copied styles that we need for richtext to content-editor.css, which
is designated for CSS that we want to apply to content.

WOOPLUG-3951

* Cleanup private APIs

WOOPLUG-3951

* Fix JS lint issues

WOOPLUG-3951

* Cleanup editore store

WOOPLUG-3951

* Add changelogs

WOOPLUG-3951

* Optimize fetching of emailPosts

We need email posts only for the template select modal for the new email.
When we fetch previews only for swapping templates, we can fetch sent emails.

WOOPLUG-3951

* Hide "Preferences" item in the more menu

Currently we don't have any preferences to display.
When we fully migrate to the post editor we need to decide
whether we use the preferences from the post editor or keep
the memu item hidden.

WOOPLUG-3951

* Update the snackbar notice when updating the email post.

WOOPLUG-3951

* Update preview dropdown item text for sending preview

The new text is the same as the one used in the version before the migration.
WOOPLUG-3951

* Improve memoization in use-email-styles hook

WOOPLUG-3951

* Fixes and improvements in the publish-save hack

- fixes re-render issues by using proper selectors and moving the validation hook to SendButton component
- replaces incorrectly used useState with useRef to hold a reference of a DOM node
- add check if publish button node is present before performing any changes on that

WOOPLUG-3951

* Add error boundary for extenders of woocommerce_email_editor_template_sections filter

I chose to add the boundary per section so that the only the problematic sections
so that we can still render sections that work.
WOOPLUG-3951

* Cleanup types enhacements for wordpress/interface

WOOPLUG-3951

* Fix the style loading for the editor user

WOOPLUG-3951

* Address types and linter errors

WOOPLUG-3951

* Reintroduce block compatibility warnings in the sidebar

WOOPLUG-3951