UI improvements and style panel fixes in the email editor (#58078)
* Update texts in preview modal

WOOPLUG-4200

* Address deprecation warnings from the TextControl in the preview modal

- opt-in to the new default input height which matches the design
- opt-in to the new default no bottom margin. This also required adding
the margin using CSS to match the design.

WOOPLUG-4200

* Update switch to template mode modals

The modal opened by clicking on canvas is handled by core, so I chose to use translation
hooks to replace the copy.
WOOPLUG-4200

* Fix the styles preview component to reflect changes

The main issue was that the Preview component was using static settings
value when reading style, which was set by the server.

I updated it to use live styles from the useEmailStyles hook, but I had to do additional
changes to be able to get values when the style was set as CSS variable.

WOOPLUG-4200

* Fix that the theme color palette was not available in style > color settings

The color selection component uses theme settings to generate color palettes.
The site palette was added only to editor settings but was not present in the base theme,
which is used for the color selection component in the ScreenColors component.

This commit moves the logic for extracting handpicked site theme settings and styles
to a proper place.

WOOPLUG-4200

* Remove hardcoded heading color default

This improves alignment with the site's global styles.
Heading color can be unset, and the heading will use the text color.
In case there is an extra color set for the headings, it is set
by WordPress's style system on the nested h1 element.

This also fixes an issue with Outlook, which was ignoring the color
set on the main wrapper.

WOOPLUG-4200

* Add changelogs

WOOPLUG-4200

* Remove duplicate declaration for EmailStyles

WOOPLUG-4200

* Update preview component highlight colors selection logic

This commit updates the logic for selection of highlighted
colors in the preview component to be aligned with the updated
logic in the site editor.

WOOPLUG-4200