Wooplug 4464 email editor crashes (#58459)
* Fix occasional crash when switching email editor to template mode

I found that during the switch to template mode, there is a short period of time
when the component is rendered, but the post in the editor store is already switched
to the template.

This commit fixes the crash by not rendering components when data are not available.
WOOPLUG-4464

* Wrap sidebar component added via a filter to the Error boundary

This prevents errors from components extending sidebar to crash
the whole editor

WOOPLUG-4464

* Add change logs