Update address JS utils & remove filter (#58030)
* Update address JS utils to use the addressFieldsForShippingRates setting
Refactored and renamed the isAddressComplete function to use the shipping fields setting; we're not using it for checking random fields, it was used only for the fields from addressFieldsForShippingRates
Removed some JS utils we weren't using

* Fix typo

* add changelog

* Add the addressFieldsForShippingRates setting to cart

* Remove the woocommerce_address_fields_for_shipping_rates filter

This was an optimistically added filter. We're not accounting for it in the shipping calculator and the shipping logic isn't account for additional shipping fields

* Remove addressFieldsForShippingRates from settings
This is a JS only constant, so there is no need to store it in the settings via the assets registry

* Fix type and lint error

ResponseBaseAddress relates to Store API types so I'm changing the type base.