Fix outdated content when using Preview in new tab in the block email editor (#58481)
* Add updatePreviewInNewTabLink functionality to enhance preview behavior

Why: The default editor preview link has a preview_nonce added to all  published posts. This caches the post content and prevent the user from seeing the latest content update in the preview tab.

* Introduced a new module to manage the preview in new tab link, ensuring it updates correctly and prevents caching issues.
* Integrated the updatePreviewInNewTabLink function into the Editor initialization process to enhance user experience.
* Added event listeners to handle click events and track user interactions for analytics.

This change improves the reliability of the preview feature in the email editor.

* Add changelog file

* Refactor email editor preview functionality

Removed the outdated updatePreviewInNewTabLink function and its associated module to streamline the email editor's preview behavior. Introduced a new method to update the preview post link by removing the preview_nonce, ensuring users see the latest content in the preview. Updated the changelog to reflect this enhancement.