Order summary: Switch to css container width detection for the cart and checkout block (#58476)
* Switch to css container width detection for the cart and checkout block

Using JS to detect container width impacts the initial rendering of the blocks adding unnecessary layout shifts.
These changes only account for the order summary area.
The is-large, is-small etc classes are kept for backwards compatibility.

* Add changelog

* Fix country select in the editor

* Rename css mixin

* Fix duplication

* Revert to JS handling of the CheckoutOrderSummaryFill
Using the reverted condition, to make sure this doesn't get rendered before we have a value for !isLarge

* Add comment for clarity