[dev] Monorepo: consolidate packages license versions to `GPL-2.0-or-later` (#58941)
Currently, we have a mix of GPL-compatible licenses specified in different formats due to historical reasons and the merging of separate repositories into a single monorepo over a short period. The GPL-2.0-or-later specifier covers already used specifiers, and is aligned with  https://developer.wordpress.org/plugins/wordpress-org/detailed-plugin-guidelines/#1-plugins-must-be-compatible-with-the-gnu-general-public-license and other monorepos ((Gutenberg, Calypso, Jetpack)) - hence applied to all packages in this repository.