Add the usage of the render_email_callback to the email rendering. (#59070)
* Use the custom block property render_email_callback for rendering

This change allows defining a rendering callback that is used for rendering in emails.

* Remove the class Blocks_Registry and its test

The class Blocks_Registry is not needed anymore when we use the block custom callback.

* Replace using Blocks_Registry with register_block_type function

- I didn't move placeholder const to avoid changes in the email templates.

* Add changelogs

* Remove the list of allowed blocks from email-editor JS package

We decided to keep rendering callbacks and block configuration at one place. This place should be PHP package.

* Add settings supports.email to true for allowed core blocks

* Add changelog and update readme.md

* Refactor setting render_email_callback to one method

This refactoring should prevent the creation of class instances earlier than necessary.

* Add internal property for caching renderer instances

* Add try-catch statement around calling the custom callback

* Fix the incorrect exit and change method to private