Update Checkout block initial render loading (#58782)
* Switch to css container width detection for the cart and checkout block

Using JS to detect container width impacts the initial rendering of the blocks adding unnecessary layout shifts.
These changes only account for the order summary area.
The is-large, is-small etc classes are kept for backwards compatibility.

* Add changelog

* Fix country select in the editor

* Remove Suspense from Checkout block rendering

* Fix content jump for selects as first inputs while improving css code for it

* Fix content jump for order summary and address card

* Add loading states to Order summary
Also fixed product variation styling, which was meant initially for the order summary of cart

* Fix content jump on the order summary

* Add Payment options skeleton

* Improve payment options skeleton

* Add loading state for the shipping options

* Remove animation of Address Card on initial load
Logic for expanded/collapsed state on first render will be refactored in another PR

* Improve skeleton size for payment/shipping options

* Refactor size based styles to use the mixin

* Show billing address form on first render

* Show shipping form on first load

* Temporary improvement of the initial skeleton

* Fix typo

* Rename css mixin for consistency

* Fix mixin name

* Fix storybook

* Use container width mixins for checkout block content

* Add changelog

* Fix skeleton

* Fix linting

* Fix js test

* Fix css property

* Add a11y to skeleton components

* Fix test mocks

* Update test with assertions instead of snapshots

* Remove lazy loading for checkout inner blocks
Having inner blocks lazy loaded is a blocker for removing suspense.
After some analysis, the impact on the bundle size is minimal. With this approach we're actually improving the speed score for the page, from 1.7 to 1.1s.
Moving forward, we could consider a hybrid approach if we ever move to a stepped checkout

* Remove conditional for block loading for the parent checkout block

* Fix linting

* Fix suspense loading for inner components

* Add back key for suspense

* Fix isAddingToCart to support testing environments
With the checkout now not being lazy loaded, the cart store is initialized in the testing env.
Meaning that tests that initialize the window object would trigger this calculation

* Fix dependency loop by removing the Skeleton import from the TotalItem component and moving it above

* Fix order summary spacing issue

* Fix type until loading is added to the cart block as well

* Fix mobile small styles for payment method label

* Import Skeleton in TotalsItem directly

* Fix SCSS include

* Rename animation class

* Make screen reader loading texts more user friendly

* Fix desktop structure of the Order summary title
The title had a bug where the price and the accordion arrow were show, which are elements specific for smaller layouts

* Try to click checkbox in e2e test for more stable results