Refactor reading editor data from global window object (#59186)
* Update initial state to use current_post_id and current_post_type from WooCommerceEmailEditor

* Refactor email editor selectors to use dynamic post type retrieval

Updated the email editor selectors to replace the hardcoded `editorCurrentPostType` with a dynamic retrieval of the post type from the store. This change enhances flexibility and ensures that the selectors work correctly with varying post types.

* Remove unused event tracking function and add postType to state type in email editor.

We don't need this tracking anymore because it is done in `packages/js/email-editor/src/components/header/trash-email-post.tsx`

* Refactor email editor to retrieve post type dynamically from store

Updated the email editor component to use the dynamically retrieved post type from the store instead of the previously hardcoded value. This change enhances the flexibility of the editor and aligns with recent improvements in state management.

* Enhance template selection to utilize dynamic post type

I chose the prop-drilling approach because the component is already quite complex, and I want to avoid impacting its performance by introducing another unnecessary state.

Updated the TemplateSelection and SelectTemplateModal components to pass the dynamically retrieved post type from the store. This change improves the flexibility of template selection and ensures compatibility with various post types.

* Remove hardcoded post type and ID from constants in email editor store. This change aligns with recent updates to dynamically retrieve these values, enhancing flexibility and maintainability.

* Add changelog file

* Fix typos in telemetry tracking documentation and improve error handling in initial state setup. Added checks for required properties from the WooCommerceEmailEditor global object to ensure proper functionality of the email editor.

* Update email preview test to include dynamic email post type retrieval in setup mock. This change enhances the test's accuracy by reflecting the latest state management improvements.